requirements.txt
asgiref==3.5.2
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
numpy==1.21.6
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
ptyprocess==0.7.0
Pygments==2.13.0
python-dateutil==2.8.2
pytrends==4.8.0
pytz==2022.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.9.0
